+++ tests/functional/functional_grpc_plugins_test.rs
@@ -222,7 +222,7 @@ async fn send_grpc_request(
 
     let mut req_builder = Request::builder()
         .method("POST")
-        .uri(path)
+        .uri(format!("http://{addr}{path}"))
         .header("content-type", "application/grpc")
         .header("te", "trailers");